fix: segregate error classes to fix circular dependancy
