+++ src/errors/index.ts
@@ -1,625 +1,32 @@
-export abstract class SdkError extends Error {
-  public abstract code: string;
-}
+export { SdkError } from "./sdk-error.js";
 
-/**
- * Errors that come from Auth0 in the `redirect_uri` callback may contain reflected user input via the OpenID Connect `error` and `error_description` query parameter.
- * You should **not** render the error `message`, or `error` and `error_description` properties without properly escaping them first.
- */
-export class OAuth2Error extends SdkError {
-  public code: string;
+export {
+  OAuth2Error,
+  DiscoveryError,
+  MissingStateError,
+  InvalidStateError,
+  InvalidConfigurationError,
+  AuthorizationError,
+  AuthorizationCodeGrantRequestError,
+  AuthorizationCodeGrantError,
+  BackchannelLogoutError,
+  BackchannelAuthenticationNotSupportedError,
+  BackchannelAuthenticationError,
+  AccessTokenError,
+  AccessTokenErrorCode,
+  AccessTokenForConnectionError,
+  AccessTokenForConnectionErrorCode,
+  CustomTokenExchangeError,
+  CustomTokenExchangeErrorCode
+} from "./oauth-errors.js";
 
-  constructor({ code, message }: { code: string; message?: string }) {
-    super(
-      message ??
-        "An error occurred while interacting with the authorization server."
-    );
-    this.name = "OAuth2Error";
-    this.code = code;
-  }
-}
+export { DPoPError, DPoPErrorCode } from "./dpop-errors.js";
 
-export class DiscoveryError extends SdkError {
-  public code: string = "discovery_error";
-
-  constructor(message?: string) {
-    super(message ?? "Discovery failed for the OpenID Connect configuration.");
-    this.name = "DiscoveryError";
-  }
-}
-
-export class MissingStateError extends SdkError {
-  public code: string = "missing_state";
-
-  constructor(message?: string) {
-    super(message ?? "The state parameter is missing.");
-    this.name = "MissingStateError";
-  }
-}
-
-export class InvalidStateError extends SdkError {
-  public code: string = "invalid_state";
-
-  constructor(message?: string) {
-    super(message ?? "The state parameter is invalid.");
-    this.name = "InvalidStateError";
-  }
-}
-
-export class InvalidConfigurationError extends SdkError {
-  public code: string = "invalid_configuration";
-
-  constructor(message?: string) {
-    super(message ?? "The configuration is invalid.");
-    this.name = "InvalidConfigurationError";
-  }
-}
-
-export class AuthorizationError extends SdkError {
-  public code: string = "authorization_error";
-  public cause: OAuth2Error;
-
-  constructor({ cause, message }: { cause: OAuth2Error; message?: string }) {
-    super(message ?? "An error occurred during the authorization flow.");
-    this.cause = cause;
-    this.name = "AuthorizationError";
-  }
-}
-
-export class AuthorizationCodeGrantRequestError extends SdkError {
-  public code: string = "authorization_code_grant_request_error";
-
-  constructor(message?: string) {
-    super(
-      message ??
-        "An error occurred while preparing or performing the authorization code grant request."
-    );
-    this.name = "AuthorizationCodeGrantRequestError";
-  }
-}
-
-export class AuthorizationCodeGrantError extends SdkError {
-  public code: string = "authorization_code_grant_error";
-  public cause: OAuth2Error;
-
-  constructor({ cause, message }: { cause: OAuth2Error; message?: string }) {
-    super(
-      message ??
-        "An error occurred while trying to exchange the authorization code."
-    );
-    this.cause = cause;
-    this.name = "AuthorizationCodeGrantError";
-  }
-}
-
-export class BackchannelLogoutError extends SdkError {
-  public code: string = "backchannel_logout_error";
-
-  constructor(message?: string) {
-    super(
-      message ??
-        "An error occurred while completing the backchannel logout request."
-    );
-    this.name = "BackchannelLogoutError";
-  }
-}
-
-export class BackchannelAuthenticationNotSupportedError extends SdkError {
-  public code: string = "backchannel_authentication_not_supported_error";
-
-  constructor() {
-    super(
-      "The authorization server does not support backchannel authentication. Learn how to enable it here: https://auth0.com/docs/get-started/applications/configure-client-initiated-backchannel-authentication"
-    );
-    this.name = "BackchannelAuthenticationNotSupportedError";
-  }
-}
-
-export class BackchannelAuthenticationError extends SdkError {
-  public code: string = "backchannel_authentication_error";
-  public cause?: OAuth2Error;
-
-  constructor({ cause }: { cause?: OAuth2Error }) {
-    super(
-      "There was an error when trying to use Client-Initiated Backchannel Authentication."
-    );
-    this.cause = cause;
-    this.name = "BackchannelAuthenticationError";
-  }
-}
-
-export enum AccessTokenErrorCode {
-  MISSING_SESSION = "missing_session",
-  MISSING_REFRESH_TOKEN = "missing_refresh_token",
-  FAILED_TO_REFRESH_TOKEN = "failed_to_refresh_token"
-}
-
-export class AccessTokenError extends SdkError {
-  public code: string;
-  public cause?: OAuth2Error;
-
-  constructor(code: string, message: string, cause?: OAuth2Error) {
-    super(message);
-    this.name = "AccessTokenError";
-    this.code = code;
-    this.cause = cause;
-  }
-}
-
-/**
- * Enum representing error codes related to access tokens for connections.
- */
-export enum AccessTokenForConnectionErrorCode {
-  /**
-   * The session is missing.
-   */
-  MISSING_SESSION = "missing_session",
-
-  /**
-   * The refresh token is missing.
-   */
-  MISSING_REFRESH_TOKEN = "missing_refresh_token",
-
-  /**
-   * Failed to exchange the refresh token.
-   */
-  FAILED_TO_EXCHANGE = "failed_to_exchange_refresh_token"
-}
-
-/**
- * Error class representing an access token for connection error.
- * Extends the `SdkError` class.
- */
-export class AccessTokenForConnectionError extends SdkError {
-  /**
-   * The error code associated with the access token error.
-   */
-  public code: string;
-  public cause?: OAuth2Error;
-
-  /**
-   * Constructs a new `AccessTokenForConnectionError` instance.
-   *
-   * @param code - The error code.
-   * @param message - The error message.
-   * @param cause - The OAuth2 cause of the error.
-   */
-  constructor(code: string, message: string, cause?: OAuth2Error) {
-    super(message);
-    this.name = "AccessTokenForConnectionError";
-    this.code = code;
-    this.cause = cause;
-  }
-}
-
-/**
- * Error codes for Custom Token Exchange errors.
- */
-export enum CustomTokenExchangeErrorCode {
-  /**
-   * The subject_token is missing or empty.
-   */
-  MISSING_SUBJECT_TOKEN = "missing_subject_token",
-
-  /**
-   * The subject_token_type is not a valid URI, wrong length, or uses a reserved namespace.
-   */
-  INVALID_SUBJECT_TOKEN_TYPE = "invalid_subject_token_type",
-
-  /**
-   * The actor_token was provided without actor_token_type.
-   */
-  MISSING_ACTOR_TOKEN_TYPE = "missing_actor_token_type",
-
-  /**
-   * The token exchange request failed.
-   */
-  EXCHANGE_FAILED = "exchange_failed"
-}
-
-/**
- * Error class representing a Custom Token Exchange error.
- * Extends the `SdkError` class.
- *
- * This error is thrown when a Custom Token Exchange operation fails,
- * such as validation errors or server-side token exchange failures.
- *
- * @see {@link https://auth0.com/docs/authenticate/custom-token-exchange Auth0 Custom Token Exchange Documentation}
- */
-export class CustomTokenExchangeError extends SdkError {
-  /**
-   * The error code associated with the custom token exchange error.
-   */
-  public code: string;
-  /**
-   * The underlying OAuth2 error that caused this error (if applicable).
-   */
-  public cause?: OAuth2Error;
-
-  /**
-   * Constructs a new `CustomTokenExchangeError` instance.
-   *
-   * @param code - The error code.
-   * @param message - The error message.
-   * @param cause - The OAuth2 cause of the error.
-   */
-  constructor(code: string, message: string, cause?: OAuth2Error) {
-    super(message);
-    this.name = "CustomTokenExchangeError";
-    this.code = code;
-    this.cause = cause;
-  }
-}
-
-/**
- * Error codes for DPoP-related errors.
- *
- * These error codes categorize different types of failures that can occur
- * during DPoP (Demonstrating Proof-of-Possession) operations.
- */
-export enum DPoPErrorCode {
-  /**
-   * Failed to calculate dpop_jkt (JWK thumbprint) parameter.
-   * This occurs when the SDK cannot generate the required thumbprint
-   * from the provided public key for the authorization request.
-   */
-  DPOP_JKT_CALCULATION_FAILED = "dpop_jkt_calculation_failed",
-
-  /**
-   * Failed to export DPoP public key to JWK format.
-   * This occurs when the SDK cannot convert the CryptoKey to the
-   * JSON Web Key format required for DPoP proofs.
-   */
-  DPOP_KEY_EXPORT_FAILED = "dpop_key_export_failed",
-
-  /**
-   * DPoP configuration is invalid or incomplete.
-   * This occurs when the provided DPoP configuration contains
-   * invalid values or missing required components.
-   */
-  DPOP_CONFIGURATION_ERROR = "dpop_configuration_error"
-}
-
-/**
- * Represents an error that occurred during DPoP (Demonstrating Proof-of-Possession) operations.
- *
- * DPoP is an OAuth 2.0 extension that provides application-level proof-of-possession security
- * by binding access tokens to cryptographic key pairs. This error is thrown when DPoP-related
- * operations fail, such as key pair operations, proof generation, or configuration issues.
- *
- * Common scenarios that trigger DPoPError:
- * - Invalid or incompatible key pairs (wrong algorithm, corrupted keys)
- * - JWK thumbprint calculation failures
- * - Public key export failures to JWK format
- * - Invalid DPoP configuration parameters
- *
- * @example Handling DPoP errors
- * ```typescript
- * try {
- *   const auth0 = new Auth0Client({
- *     useDPoP: true,
- *     dpopKeyPair: invalidKeyPair
- *   });
- * } catch (error) {
- *   if (error instanceof DPoPError) {
- *     console.error(`DPoP Error [${error.code}]:`, error.message);
- *
- *     switch (error.code) {
- *       case DPoPErrorCode.DPOP_KEY_EXPORT_FAILED:
- *         console.error("Key export failed. Check key format and algorithm.");
- *         break;
- *       case DPoPErrorCode.DPOP_JKT_CALCULATION_FAILED:
- *         console.error("JWK thumbprint calculation failed. Verify key validity.");
- *         break;
- *       case DPoPErrorCode.DPOP_CONFIGURATION_ERROR:
- *         console.error("Invalid DPoP configuration. Check options and environment variables.");
- *         break;
- *     }
- *
- *     if (error.cause) {
- *       console.error("Underlying cause:", error.cause);
- *     }
- *   }
- * }
- * ```
- *
- * @example Creating a DPoP error
- * ```typescript
- * throw new DPoPError(
- *   DPoPErrorCode.DPOP_JKT_CALCULATION_FAILED,
- *   "Failed to calculate dpop_jkt parameter from public key",
- *   originalError
- * );
- * ```
- *
- * @see {@link https://datatracker.ietf.org/doc/html/rfc9449 | RFC 9449: OAuth 2.0 Demonstrating Proof-of-Possession at the Application Layer (DPoP)}
- */
-export class DPoPError extends SdkError {
-  /** The specific DPoP error code indicating the type of failure */
-  public code: DPoPErrorCode;
-  /** The underlying error that caused this DPoP error (optional) */
-  public cause?: Error;
-
-  /**
-   * Constructs a new `DPoPError` instance.
-   *
-   * @param code - The DPoP error code indicating the specific type of failure
-   * @param message - A descriptive error message explaining what went wrong
-   * @param cause - The underlying error that caused this DPoP error (optional)
-   *
-   * @example
-   * ```typescript
-   * const dpopError = new DPoPError(
-   *   DPoPErrorCode.DPOP_KEY_EXPORT_FAILED,
-   *   "Unable to export public key to JWK format",
-   *   keyExportError
-   * );
-   * ```
-   */
-  constructor(code: DPoPErrorCode, message: string, cause?: Error) {
-    super(message);
-    this.name = "DPoPError";
-    this.code = code;
-    this.cause = cause;
-  }
-}
-
-/**
- * Error class representing a connect account request error.
- */
-export class MyAccountApiError extends SdkError {
-  public name: string = "MyAccountApiError";
-  public code: string = "my_account_api_error";
-  public type: string;
-  public title: string;
-  public detail: string;
-  public status: number;
-  public validationErrors?: Array<{
-    /**
-     * A human-readable description of the specific error. Required.
-     */
-    detail: string;
-    /**
-     * The name of the invalid parameter. Optional.
-     */
-    field?: string;
-    /**
-     * A JSON Pointer that points to the exact location of the error in a JSON document being validated. Optional.
-     */
-    pointer?: string;
-    /**
-     *  Specifies the source of the error (e.g., body, query, or header in an HTML message). Optional.
-     */
-    source?: string;
-  }>;
-
-  constructor({
-    type,
-    title,
-    detail,
-    status,
-    validationErrors
-  }: {
-    type: string;
-    title: string;
-    detail: string;
-    status: number;
-    validationErrors?: Array<{
-      detail: string;
-      field?: string;
-      pointer?: string;
-      source?: string;
-    }>;
-  }) {
-    super(`${title}: ${detail}`);
-    this.type = type;
-    this.title = title;
-    this.detail = detail;
-    this.status = status;
-    this.validationErrors = validationErrors;
-  }
-}
-
-/**
- * Enum representing error codes related to the connect account flow.
- */
-export enum ConnectAccountErrorCodes {
-  /**
-   * The session is missing.
-   */
-  MISSING_SESSION = "missing_session",
-
-  /**
-   * Failed to initiate the connect account flow.
-   */
-  FAILED_TO_INITIATE = "failed_to_initiate",
-
-  /**
-   * Failed to complete the connect account flow.
-   */
-  FAILED_TO_COMPLETE = "failed_to_complete"
-}
-
-/**
- * Error class representing a connect account error.
- */
-export class ConnectAccountError extends SdkError {
-  /**
-   * The error code associated with the connect account error.
-   */
-  public code: string;
-  public cause?: MyAccountApiError;
-
-  constructor({
-    code,
-    message,
-    cause
-  }: {
-    code: string;
-    message: string;
-    cause?: MyAccountApiError;
-  }) {
-    super(message);
-    this.name = "ConnectAccountError";
-    this.code = code;
-    this.cause = cause;
-  }
-}
-
-/**
- * MFA requirements from an mfa_required error response.
- * Indicates which MFA methods are available for the user.
- * Matches Auth0 API response shape and auth0-spa-js type.
- */
-export interface MfaRequirements {
-  /** Required enrollment types (user needs to enroll new authenticator) */
-  enroll?: Array<{ type: string }>;
-  /** Available challenge types (existing authenticators) */
-  challenge?: Array<{ type: string }>;
-}
-
-/**
- * Thrown when {@link getAccessToken} requires MFA step-up authentication.
- *
- * This error is thrown during token refresh when Auth0 returns `mfa_required`.
- * The {@link mfa_token} property contains an encrypted token that can be used
- * with Auth0's MFA API to complete the authentication challenge.
- *
- * @remarks
- * The `mfa_token` is encrypted using the SDK's cookie secret for security.
- * The raw token from Auth0 is never exposed to application code.
- *
- * Supports two consumption paths with identical shape:
- * 1. Direct SDK call: properties accessed on error instance
- * 2. HTTP API route: Response.json(error) uses toJSON() automatically
- *
- * @example Handling MFA required in a route handler
- * ```typescript
- * import { getAccessToken, MfaRequiredError } from "@auth0/nextjs-auth0/server";
- *
- * try {
- *   const { token } = await getAccessToken({ audience: "https://api.example.com" });
- * } catch (error) {
- *   if (error instanceof MfaRequiredError) {
- *     // Redirect to MFA challenge page
- *     redirect(`/mfa?token=${error.mfa_token}`);
- *   }
- *   throw error;
- * }
- * ```
- *
- * @see {@link https://auth0.com/docs/api/authentication#multi-factor-authentication Auth0 MFA API}
- */
-export class MfaRequiredError extends SdkError {
-  public readonly code: string = "mfa_required";
-
-  /**
-   * Encrypted MFA token to pass to MFA API methods.
-   */
-  public readonly mfa_token: string;
-
-  /** Original Auth0 error code */
-  public readonly error: string = "mfa_required";
-
-  /** Original Auth0 error description */
-  public readonly error_description: string;
-
-  /** MFA requirements indicating available challenge/enrollment methods */
-  public readonly mfa_requirements?: MfaRequirements;
-
-  public readonly cause?: OAuth2Error;
-
-  /**
-   * @param error_description - Error description from Auth0
-   * @param mfaToken - Encrypted MFA token (constructor param uses camelCase)
-   * @param mfaRequirements - MFA requirements from Auth0 (constructor param uses camelCase)
-   * @param cause - Underlying OAuth2 error
-   *
-   * @remarks
-   * Constructor parameters use camelCase (mfaToken, mfaRequirements) for consistency
-   * with SDK conventions, but they are assigned to snake_case properties (mfa_token,
-   * mfa_requirements) to match Auth0 API response format.
-   */
-  constructor(
-    error_description: string,
-    mfaToken: string,
-    mfaRequirements?: MfaRequirements,
-    cause?: OAuth2Error
-  ) {
-    super(error_description);
-    this.name = "MfaRequiredError";
-    this.error_description = error_description;
-    this.mfa_token = mfaToken;
-    this.mfa_requirements = mfaRequirements;
-    this.cause = cause;
-  }
-
-  /**
-   * Serialize error for HTTP responses.
-   * Called automatically by Response.json() and JSON.stringify().
-   * Ensures both SDK and HTTP API consumers get identical shape.
-   */
-  toJSON(): {
-    error: string;
-    error_description: string;
-    mfa_token: string;
-    mfa_requirements?: MfaRequirements;
-  } {
-    return {
-      error: this.error,
-      error_description: this.error_description,
-      mfa_token: this.mfa_token,
-      ...(this.mfa_requirements && { mfa_requirements: this.mfa_requirements })
-    };
-  }
-}
-
-/**
- * Thrown when MFA API methods are called but no context exists in session
- * for the provided encrypted mfa_token.
- *
- * This typically occurs when:
- * - The session expired between catching MfaRequiredError and calling MFA methods
- * - The mfa_token was modified or is from a different session
- * - The MFA context was cleaned up due to TTL expiration
- *
- * @example
- * ```typescript
- * try {
- *   await auth0.completeMfaChallenge(mfaToken, code);
- * } catch (error) {
- *   if (error instanceof MfaTokenNotFoundError) {
- *     // Restart MFA flow - context was lost
- *     redirect("/auth/login?prompt=mfa");
- *   }
- * }
- * ```
- */
-export class MfaTokenExpiredError extends SdkError {
-  public readonly code: string = "mfa_token_expired";
-
-  constructor() {
-    super("MFA token has expired. Please restart the MFA flow.");
-    this.name = "MfaTokenExpiredError";
-  }
-}
-
-/**
- * Thrown when the encrypted mfa_token is invalid.
- *
- * This occurs when:
- * - The token was tampered with
- * - The token is malformed (not valid JWE)
- * - The token was encrypted with a different secret
- */
-export class MfaTokenInvalidError extends SdkError {
-  public readonly code: string = "mfa_token_invalid";
-
-  constructor() {
-    super("MFA token is invalid.");
-    this.name = "MfaTokenInvalidError";
-  }
-}
+export {
+  MyAccountApiError,
+  ConnectAccountError,
+  ConnectAccountErrorCodes
+} from "./my-account-errors.js";
 
 export {
   MfaGetAuthenticatorsError,
@@ -627,6 +34,10 @@ export {
   MfaVerifyError,
   MfaEnrollmentError,
   MfaNoAvailableFactorsError,
-  type MfaApiErrorResponse
-} from "./mfa.js";
-export { InvalidRequestError } from "./invalid-request-error.js";
+  MfaRequiredError,
+  MfaTokenExpiredError,
+  MfaTokenInvalidError,
+  InvalidRequestError,
+  type MfaApiErrorResponse,
+  type MfaRequirements
+} from "./mfa-errors.js";
