+++ src/errors/sdk-error.ts
@@ -1,0 +1,3 @@
+export abstract class SdkError extends Error {
+  public abstract code: string;
+}
